Create sqrt_bisection_method.py
In this project, I learned how to use the bisection method in mathematics to approximate square roots without having to import a module with a square root method. Knowing approximation methods like this is important in scenarios where there may not be a straightforward way to solve them, like several differential equation problems.